requirements.txt
google-generativeai==0.3.0
flask==2.3.3
flask-cors==4.0.0
schedule==1.2.0
requests==2.31.0
beautifulsoup4==4.12.2
feedparser==6.0.10
selenium==4.15.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
gunicorn==21.2.0